+++ CHANGELOG.md
@@ -6,10 +6,23 @@
 
 BREAKING CHANGES
 * Change default ports from 466xx to 266xx
+* AltBytes renamed to Memo, now a string, max 100 characters, costs a bit of gas
+* Transactions now take a list of Messages
+* Signers of a transaction now only sign over their account and sequence number
+
+FEATURES
+* [gaiacli] You can now attach a simple text-only memo to any transaction, with the `--memo` flag
+* [lcd] Queried TXs now include the tx hash to identify each tx
+* [mockapp] CompleteSetup() no longer takes a testing parameter
+* [governance] Implemented MVP
+  * Supported proposal types: just binary (pass/fail) TextProposals for now
+  * Proposals need deposits to be votable; deposits are burned if proposal fails
+  * Delegators delegate votes to validator by default but can override (for their stake)
 
 FIXES
 * \#1259 - fix bug where certain tests that could have a nil pointer in defer
 * Retry on HTTP request failure in CLI tests, add option to retry tests in Makefile
+* Fixed bug where chain ID wasn't passed properly in x/bank REST handler
 
 ## 0.19.0